ci(test): refactor for splitting test to multiple files
