+++ test/WebpackNotifierPlugin.test.ts
@@ -1,7 +1,4 @@
-import {notify} from 'node-notifier';
-import WebpackNotifierPlugin from '../';
-import {getCompiler, compile, prepareFs, contentImageSerializer, reduceArraySerializer} from './utils';
-import fixtures from "./fixtures";
+import {contentImageSerializer, reduceArraySerializer, testChangesFlow} from './utils';
 
 expect.addSnapshotSerializer(reduceArraySerializer);
 expect.addSnapshotSerializer(contentImageSerializer);
@@ -43,16 +40,3 @@ describe('WebpackNotifierPlugin', () => {
     });
   });
 });
-
-async function testChangesFlow(sources, opts)  {
-  const compiler = getCompiler();
-  const plugin = new WebpackNotifierPlugin(opts);
-  plugin.apply(compiler);
-
-  for (const name of sources) {
-    notify.mockClear();
-    prepareFs(fixtures.simple[name]);
-    await compile(compiler);
-    expect(notify.mock.calls).toMatchSnapshot(`after "${name}" build`);
-  }
-}
